Maybe make context calculation faster
It bothered me, that we would iterate the view twice; with expensive set.constains calls.
Once to find its size, and again to copy everything.
We still iterate twice, because it makes sense to resize the list,
but in this implementation the iterations are as cheap as they get.
Maybe we even get some cache preloading. Could be by the looks of it.

Micro-benchmark are meaningless in gc environments.
Good benchmarks are hard. Let's just trust my instinct :)